Deduplicate component/link materialization between InMemoryProjection and MongoProjection
MaterializeLink, MaterializeComponent, ResolveComponent, GetComponentTypeId, and
GetComponentId were copy-pasted near-verbatim between the two projections (~150 lines).
None of them touch projection-specific state — they're pure functions over Type/object?/
IComponentContainer/SingleObjectEvent/ITypeMetadataProvider — so they move to
ComponentApplyHelpers/LinkApplyHelpers in Synqra.Projection.CommonStoreSupport, which
both projections already reference. ResolveContainer stays a thin per-projection wrapper,
since the actual object lookup differs (InMemoryProjection's live model graph vs.
MongoProjection's per-process tracked-id map) — only the post-lookup validation
(null/IComponentContainer checks) is shared.
